remove the concept of components in bodyParser
