--- server/model/component.js
@@ -1,19 +1,0 @@
-type ComponentType =
-  | 'imageGallery'
-  | 'picture'
-  | 'video'
-  | 'list'
-  | 'tweet'
-  | 'heading';
-
-type ComponentWeight =
-  | 'default'
-  | 'leading'
-  | 'standalone'
-  | 'supporting';
-
-export type Component<T> = {|
-  weight: ComponentWeight;
-  type: ComponentType;
-  value: T;
-|}
