+++ server/model/heading.js
@@ -1,0 +1,5 @@
+type Heading = {|
+  level: number;
+  value: any; // TODO: Make this not `any`
+|}
+export function createHeading(data: Heading) { return (data: Heading); }
