+++ server/util/body-parser.js
@@ -7,19 +7,8 @@ import {createPicture} from '../model/picture';
 import {createVideo} from '../model/video';
 import {createList} from '../model/list';
 import {createTweet} from '../model/tweet';
-
-type BodyPart = {|
-  weight: string;
-  type: string;
-  value: any; // TODO: Make this not `any`
-|}
-function bodyPart(data: BodyPart) { return (data: BodyPart); }
-
-type Heading = {|
-  level: number;
-  value: any; // TODO: Make this not `any`
-|}
-function heading(data: Heading) { return (data: Heading); }
+import {createBodyPart} from '../model/body-part';
+import {createHeading} from '../model/heading';
 
 export function bodyParser(bodyText) {
   const fragment = getFragment(bodyText);
@@ -78,7 +67,7 @@ function decodeHtmlEntities(nodes) {
 }
 
 function convertWpStandfirst(node) {
-  return bodyPart({
+  return createBodyPart({
     type: 'standfirst',
     value: serializeAndCleanNode(node)
   });
@@ -89,9 +78,9 @@ export function convertWpHeading(node) {
   const isWpHeading = Boolean(headingMatch);
 
   if (isWpHeading) {
-    return bodyPart({
+    return createBodyPart({
       type: 'heading',
-      value: heading({
+      value: createHeading({
         level: headingMatch[1],
         value: serializeAndCleanNode(node.childNodes[0])
       })
@@ -115,7 +104,7 @@ export function convertWpImage(node) {
     const weightKey = Object.keys(weights).find(wpClassName => className.indexOf(wpClassName) !== -1);
     const weight = weightKey ? weights[weightKey] : 'default';
 
-    return bodyPart({
+    return createBodyPart({
       weight,
       type: 'picture',
       value: picture
@@ -145,7 +134,7 @@ export function convertWpVideo(node) {
     const embedUrl = getAttrVal(iframe.attrs, 'src');
     const video = createVideo({ embedUrl });
 
-    return bodyPart({
+    return createBodyPart({
       type: 'video',
       value: video
     });
@@ -168,7 +157,7 @@ export function convertWpList(node) {
       return itemVal;
     });
 
-    return bodyPart({
+    return createBodyPart({
       type: 'list',
       value: createList({
         // TODO: We should be sending a name with all lists
@@ -186,7 +175,7 @@ function convertTweet(node) {
   const isTweet = Boolean(className && className.match('embed-twitter'));
 
   if (isTweet) {
-    return bodyPart({
+    return createBodyPart({
       type: 'tweet',
       value: createTweet({
         html: serializeNode(node)
@@ -222,7 +211,7 @@ export function findWpImageGallery(node) {
           });
         });
 
-      return bodyPart({
+      return createBodyPart({
         type: 'imageGallery',
         weight: 'standalone',
         value: createImageGallery({
@@ -265,7 +254,7 @@ export function convertDomNode(node) {
   const cleanedNode = serializeAndCleanNode(node);
 
   if (cleanedNode) {
-    return bodyPart({
+    return createBodyPart({
       type: 'html',
       value: cleanedNode
     });
